requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
antlr4-python3-runtime=4.9.3=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py39h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.9.24=h06a4308_0
certifi=2024.8.30=py39h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cuda-version=12.6=3
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py39h06a4308_0
freetype=2.12.1=h4a9f257_0
fvcore=0.1.5.post20221221=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
idna=3.7=py39h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
iopath=0.1.10=pypi_0
jinja2=3.1.4=py39h06a4308_0
jpeg=9e=h5eee18b_3
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.11.1.6=0
libcurand=10.3.7.77=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py39h5eee18b_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.10=py39h5eee18b_0
mkl_random=1.2.7=py39h1128e8f_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=py39h06a4308_0
numpy=1.22.4=pypi_0
omegaconf=2.3.0=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.0.15=h5eee18b_0
packaging=24.1=pypi_0
pillow=10.4.0=py39h5eee18b_0
pip=24.2=py39h06a4308_0
portalocker=2.10.1=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.20=he870216_1
pytorch=2.0.0=py3.9_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=py39h06a4308_0
setuptools=75.1.0=py39h06a4308_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.2=py39h06a4308_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
termcolor=2.5.0=pypi_0
tk=8.6.14=h39e8969_0
torchaudio=2.0.0=py39_cu117
torchmetrics=0.10.3=pypi_0
torchtriton=2.0.0=py39
torchvision=0.15.0=py39_cu117
tqdm=4.66.5=py39h2f386ee_0
typing_extensions=4.11.0=py39h06a4308_0
tzdata=2024b=h04d1e81_0
urllib3=2.2.3=py39h06a4308_0
wheel=0.44.0=py39h06a4308_0
xz=5.4.6=h5eee18b_1
yacs=0.1.8=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
